+++ framework/react.ts
@@ -1,0 +1,66 @@
+export { createDialogRevealPlan, DIALOG_TYPE_MS, getDialogCompletionAdvanceMode } from './components/DialogReveal.ts'
+export type { DialogRevealPlan } from './components/DialogReveal.ts'
+export { SaveControlsBar } from './components/SaveControlsBar.tsx'
+export type { SaveControlsBarProps } from './components/SaveControlsBar.tsx'
+export { formatPlaytime, formatTimestamp, resolveSaveThumbnail, SaveLoadMenu } from './components/SaveLoadMenu.tsx'
+export { mountVnApp } from './components/VnApp.tsx'
+export type { VnAppComponents, VnAppOptions } from './components/VnApp.tsx'
+export {
+  backgroundSizeForFit,
+  objectFitForFit,
+  resolveBackgroundAsset,
+  selectBackgroundVariant,
+  VnBackground,
+} from './components/VnBackground.tsx'
+export type { BackgroundConfig, BackgroundVariant, VnBackgroundProps } from './components/VnBackground.tsx'
+export { VnBacklog } from './components/VnBacklog.tsx'
+export type { VnBacklogProps } from './components/VnBacklog.tsx'
+export { AsepriteSpritesheetRenderer, VnCharacter } from './components/VnCharacter.tsx'
+export type { CharacterData, VnCharacterProps } from './components/VnCharacter.tsx'
+export { VnChoices } from './components/VnChoices.tsx'
+export type { VnChoicesProps } from './components/VnChoices.tsx'
+export { VnDevtoolsOverlay } from './components/VnDevtoolsOverlay.tsx'
+export type { VnDevtoolsOverlayProps } from './components/VnDevtoolsOverlay.tsx'
+export { VnDialog } from './components/VnDialog.tsx'
+export type { DialogOptions, VnDialogHandle, VnDialogProps } from './components/VnDialog.tsx'
+export { effectDurationMs, VnEffectsLayer } from './components/VnEffectsLayer.tsx'
+export type { VnEffectState, VnEffectsLayerProps } from './components/VnEffectsLayer.tsx'
+export { VnEndScreen } from './components/VnEndScreen.tsx'
+export type { VnEndScreenProps } from './components/VnEndScreen.tsx'
+export { VnGallery } from './components/VnGallery.tsx'
+export type { VnGalleryProps } from './components/VnGallery.tsx'
+export {
+  DEFAULT_PLAYER_INPUT_MAP,
+  mergePlayerInputMap,
+  resolveKeyboardAction,
+} from './components/VnInputMap.ts'
+export type { PlayerInputAction, PlayerInputMap } from './components/VnInputMap.ts'
+export { VnMusicRoom } from './components/VnMusicRoom.tsx'
+export type { VnMusicRoomProps } from './components/VnMusicRoom.tsx'
+export { getAutoDelayMs, getAutoModeAction, getSkipModeAction } from './components/VnPlayerModes.ts'
+export type { PlayerAdvanceModeAction, PlayerAdvanceModeState } from './components/VnPlayerModes.ts'
+export { quickSave, VnQuickSave } from './components/VnQuickSave.tsx'
+export type { VnQuickSaveProps } from './components/VnQuickSave.tsx'
+export { VnSaveMenu } from './components/VnSaveMenu.tsx'
+export type { VnSaveMenuProps } from './components/VnSaveMenu.tsx'
+export { VnSettings } from './components/VnSettings.tsx'
+export type { VnSettingsProps } from './components/VnSettings.tsx'
+export { VnStage } from './components/VnStage.tsx'
+export type {
+  CharacterState,
+  SceneState,
+  TransitionState,
+  VnStageComponents,
+  VnStageProps,
+} from './components/VnStage.tsx'
+export {
+  getStageAdvanceAction,
+  isAcceptedAdvanceKey,
+} from './components/VnStageAdvance.ts'
+export type { StageAdvanceAction } from './components/VnStageAdvance.ts'
+export { VnStartMenu } from './components/VnStartMenu.tsx'
+export type { VnStartMenuProps } from './components/VnStartMenu.tsx'
+export { VnTransition } from './components/VnTransition.tsx'
+export type { TransitionConfig as VnTransitionConfig, VnTransitionProps } from './components/VnTransition.tsx'
+export { VnVolumeControl } from './components/VnVolumeControl.tsx'
+export type { VnVolumeControlProps } from './components/VnVolumeControl.tsx'