+++ framework/types.ts
@@ -1,0 +1,20 @@
+export type * from './types/audio.d.ts'
+export type * from './types/characters.d.ts'
+export type * from './types/diagnostics.d.ts'
+export type * from './types/events.d.ts'
+export type * from './types/extras.d.ts'
+export type * from './types/game-config.d.ts'
+export type * from './types/minigames.d.ts'
+export type * from './types/plugins.d.ts'
+export type * from './types/save.d.ts'
+export type * from './types/scenes.d.ts'
+export type {
+  BackgroundRendererProps,
+  CharacterRendererProps,
+  ExtrasRendererProps,
+  OverlayRendererProps,
+  RendererKind,
+  RendererPropsByKind,
+  RendererRecord,
+  TransitionRendererProps,
+} from './renderers/RendererRegistry.ts'