+++ games/mi-novela/game.config.ts
@@ -1,6 +1,6 @@
 // games/mi-novela/game.config.ts
-import { officialPlugins } from '../../framework/plugins/prebuilt/index.ts'
-import type { GameConfig } from '../../framework/types/game-config.d.ts'
+import { officialPlugins } from '../../framework/plugins.ts'
+import type { GameConfig } from '../../framework/types.ts'
 
 const config: GameConfig = {